Tried adding support for multiple frameworks
Tried adding support for net451. Seems to work for application, though get warnings re restoring some dependencies.
Don't get dll for testing for net451, which makes sense, as it isn't core.
Don't think that it makes any sense to try to maintain this support, so am just saving this in case
want to go back to it, but will remove immediately.